Rename index.ts to make next commit readable
